#455 Refactor System Testing

+++ aaran/Testing/Database/Migrations/2024_06_24_000610_create_admin_tests_table.php
@@ -11,20 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('admin_tests', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('module_id')->references('id')->on('test_modules')->onDelete('cascade');
-            $table->foreignId('db_id')->references('id')->on('db_tests')->onDelete('cascade');
-            $table->string('vname');
-            $table->longText('description');
-            $table->boolean('checked_1');
-            $table->boolean('checked_2');
-            $table->boolean('checked_3');
-            $table->longText('comment');
-            $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->smallInteger('active_id');
-            $table->timestamps();
-        });
+        if (Aaran\Aadmin\Src\DbMigration::hasDeveloper()) {
+
+            Schema::create('admin_tests', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('module_id')->references('id')->on('test_modules')->onDelete('cascade');
+                $table->foreignId('db_id')->references('id')->on('db_tests')->onDelete('cascade');
+                $table->string('vname');
+                $table->longText('description');
+                $table->boolean('checked_1');
+                $table->boolean('checked_2');
+                $table->string('db_mig',3)->nullable();
+                $table->longText('comment');
+                $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->smallInteger('active_id');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
